MODPATBLK-41 create generation services
